[refactor] 이슈 필터링을 service로 분리
