+++ backend/package.json
@@ -13,7 +13,8 @@
     "@lib": "./lib",
     "@middlewares": "./middlewares",
     "@models": "./models",
-    "@routes": "./routes"
+    "@routes": "./routes",
+    "@services": "./services"
   },
   "dependencies": {
     "axios": "^0.20.0",
