+++ backend/services/index.js
@@ -1,0 +1,1 @@
+export { issueService } from './issue.service';
